refactor(golang): extract type inference into typeinfer.go
Split golang.go (952 lines) into a walk + symbol/edge-emission core and a new
typeinfer.go holding the full Go type-inference cluster: the local type map,
inferType, selector resolution, and the type-name unwrappers.

- Decompose the three over-threshold functions into named helpers, all under
  the gate: buildTypeMap -> collectReceiverTypes/collectParamTypes;
  collectRangeVars -> registerRangeLocals/rangeValueNode/rangeElemType;
  inferType -> inferFromComposite/inferFromUnary/inferFromCall.
- Cover receiverType with a white-box fixture over real value/pointer/generic
  receivers plus the nil contract (63.6 -> 72.7%); cover the single-var and
  blank-value range paths (collectRangeVars 76.9 -> 92.3%).

Behavior unchanged: the nine go goldens are byte-identical, and a digest over
5,650 GOROOT/src files is byte-identical between the pre- and post-split
binaries (an external corpus, avoiding the self-scan file_path confound).
